+++ renderer/components/StartPage.jsx
@@ -1,16 +1,14 @@
-import { useEffect, useRef, useState } from 'react'
+import { useEffect, useState } from 'react'
 
 import Image from 'next/image'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
 
-import { useCurrentUser } from '@/lib/UserContext'
 import { useTranslation } from '@/next-i18next'
 import useSupabaseClient from 'utils/supabaseClient'
 
-import ButtonLoading from './ButtonLoading'
 import LanguageSwitcher from './LanguageSwitcher'
-import InputField from './UI/InputField'
+import Login from './Login'
 
 import LevelLogo from 'public/icons/level-logo-color.svg'
 
@@ -19,41 +17,11 @@ export default function StartPage() {
   const supabase = useSupabaseClient()
   const { t } = useTranslation(['projects', 'users'])
 
-  const [isLoadingLogin, setIsLoadingLogin] = useState(false)
-  const [login, setLogin] = useState('admin@mail.com')
-  const [password, setPassword] = useState('123456')
   const [isLoginFormVisible, setIsLoginFormVisible] = useState(false)
-  const [isError, setIsError] = useState(false)
-
-  const passwordRef = useRef(null)
-  const loginRef = useRef(null)
 
   useEffect(() => {
     window.electronAPI.setItem('isNeedAutorized', false)
   }, [])
-
-  const handleLogin = async (e) => {
-    e.preventDefault()
-    setIsLoadingLogin(true)
-    try {
-      const { error } = await supabase.auth.signInWithPassword({ email: login, password })
-      if (error) throw error
-      setIsError(false)
-
-      router.push('/account')
-    } catch (error) {
-      console.log(error)
-      setIsError(true)
-    } finally {
-      setIsLoadingLogin(false)
-    }
-  }
-
-  useEffect(() => {
-    if (loginRef?.current) {
-      loginRef.current.focus()
-    }
-  }, [isLoginFormVisible])
 
   const checkAgreements = () => {
     const agreements = window.electronAPI.getItem('agreements')
@@ -68,8 +36,6 @@ export default function StartPage() {
 
   const isAuthorization = () => {
     window.electronAPI.setItem('isNeedAutorized', true)
-
-    setIsLoadingLogin(false)
     setIsLoginFormVisible(true)
   }
 
@@ -104,10 +70,7 @@ export default function StartPage() {
                 </p>
               </div>
 
-              <div
-                className="rounded-3xl bg-th-primary-100"
-                onClick={() => isAuthorization()}
-              >
+              <div className="rounded-3xl bg-th-primary-100" onClick={isAuthorization}>
                 <p className="green-two-layers cursor-pointer rounded-3xl px-7 py-8 text-xl text-th-secondary-10 after:rounded-3xl">
                   {t('users:SignIn')}
                   {t('users:ByRegistration')}
@@ -116,46 +79,7 @@ export default function StartPage() {
             </>
           )}
 
-          {isLoginFormVisible && (
-            <form className="mt-4 flex flex-col gap-4 rounded-lg bg-white p-4 shadow-md">
-              <InputField
-                refInput={loginRef}
-                type="text"
-                name="floating_email"
-                id="floating_email"
-                value={login}
-                isError={isError && !login}
-                label={t('Login')}
-                onChange={(event) => setLogin(event.target.value)}
-                className="input-base-label"
-              />
-              <InputField
-                refInput={passwordRef}
-                type="password"
-                name="floating_password"
-                id="floating_password"
-                value={password}
-                isError={isError && !password}
-                label={t('Password')}
-                onChange={(event) => setPassword(event.target.value)}
-                className="input-password"
-              />
-              <ButtonLoading
-                onClick={handleLogin}
-                isLoading={isLoadingLogin}
-                className="btn-primary"
-              >
-                {t('users:SignIn')}
-              </ButtonLoading>
-              <button
-                type="button"
-                onClick={() => setIsLoginFormVisible(false)}
-                className="text-primary mt-2 text-center text-sm underline"
-              >
-                Вернуться к выбору входа
-              </button>
-            </form>
-          )}
+          {isLoginFormVisible && <Login onClose={() => setIsLoginFormVisible(false)} />}
         </div>
 
         <Link
